pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = ""
authors = [
    {name = "vzorvalpodrugu",email = "Starkovqw8904@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "mistralai (>=1.9.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
